requirements.txt
Flask==2.3.2
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.95.0
gspread==5.10.0
pytz==2023.3
gunicorn==21.2.0
